+++ contracts/library/BridgeLib.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.17;
 
 library BridgeLib {
     struct TokenInfo {
-        uint token; // uint160(address) will overflow in the token used for starknet
+        uint256 token; // uint160(address) will overflow in the token used for starknet
         address mainnetToken;
         uint8 decimals;
     }
@@ -15,7 +15,7 @@ library BridgeLib {
         uint64 maxVerifyChallengeSourceTxSecond;
         uint64 minVerifyChallengeDestTxSecond;
         uint64 maxVerifyChallengeDestTxSecond;
-        uint nativeToken;
+        uint256 nativeToken;
         address[] spvs;
     }
 }
